Fixed an issue when writing packets.
If the connection couldn't keep up, the socket would wait for the next frame, however, if a packet was to be sent that frame it could overwrite not-sent packet data or fragments. The solution was moving the packets that are yet to be sent at the beginning of the outputPcktBufferr while adjusting the packetoffset.